### 2019-04-23 #1403 @whedon
[PRE REVIEW]: DEEPaaS API: a REST API for Machine Learning and Deep Learning models
**Submitting author:** @alvarolopez (<a href="http://orcid.org/0000-0002-0013-4602">Álvaro López García</a>)
**Repository:** <a href="https://github.com/indigo-dc/DEEPaaS" target ="_blank">https://github.com/indigo-dc/DEEPaaS</a>
**Version:** 0.3.0
