### 2019-04-23 deec298
fixing typos
