### 2019-04-24 #1403 @arfon
> 👋 @arfon - How do we do this?

This is something we have to fix manually I'm afraid. We generally do this at the point of accepting the submission.
