### 2019-05-04 #1403 @trallard
Hi @alvarolopez

Have you got any reviewers in mind for this submission? You can check the list of registered reviewers ➡️ https://bit.ly/joss-reviewers](https://bit.ly/joss-reviewers) and suggest a couple of suitable ones
