### 2019-05-06 #1403 @dangunter
Hi, I can comment on the general idea  but as I don't really do ML etc. in practice I can't comment on the value proposition of such a tool -- if that is even a desired outcome of the review.
