### 2019-06-05 df411ac
Remove leftover debug print statement
