### 2019-06-05 748b60e
Handle SIGTERM signal explicitly
DEEPaaS is based on Flask that leverages the Werkzeug simple server for
the time being. The server exits nicely on SIGINT, but not on SIGTERM.
This is not a problem when running standalone, but when we re running
within a Docker container and a 'docker stop' command is issued, Docker
sends a SIGTERM signal to the PID 1 inside the container, signal that is
not handled by Werkzeug, therefore after a timeout Docker sends SIGKILL,
resulting in the container not exiting nicely (Docker exit status 137).
