+++ deepaas/api/__init__.py
@@ -14,6 +14,7 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
+
 import flask
 import flask_restplus
 from oslo_log import log as logging
