### 2019-06-06 f71e303
Update Dockerfile for Docker container creation
