### 2019-06-06 cde61f9
tox: remove Python 3.5 testing
