+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 2.1
-envlist = py{35,36,27},pep8,pip-missing-reqs
+envlist = py{36,27},pep8,pip-missing-reqs
 skipsdist = True
 
 [testenv]
