### 2019-06-06 0fa81b6
Rename "openwisk" to "openwhisk" module
