### 2019-06-06 0ddb1e5
Add unit tests for cmd module
