### 2019-06-06 2e00f1e
Update version on Dockerfile
