### 2019-06-10 #1403 @trallard
Thanks for your patience @alvarolopez let's try and find you another reviewer to move this forward

Hi @nilesh-patil are you willing and able to take on this review?
Hi @rasbst are you willing and able to take on this review?
Hi @ssomani are you willing and able to take on this review?
Hi @GregaVrbancic are you willing and able to take on this review?

**Title**: DEEPaaS
**Description**: DEEP as a Service (DEEPaaS) is a REST API that is focused on providing access to machine learning models. By using DEEPaaS users can easily run a REST API in front of their model, thus accessing its functionality via HTTP calls.
**Repo**: [https://github.com/indigo-dc/DEEPaaS](https://github.com/indigo-dc/DEEPaaS)
