### 2019-06-21 #1403 @trallard
Thanks @krother  and @rougier  I will assign you both to the review process and take it from there.

@rougier  feel free to jump into as and when you can
