### 2019-06-21 #1517 @whedon
[REVIEW]: DEEPaaS API: a REST API for Machine Learning and Deep Learning models
**Submitting author:** @alvarolopez (<a href="http://orcid.org/0000-0002-0013-4602">Álvaro López García</a>)
**Repository:** <a href="https://github.com/indigo-dc/DEEPaaS" target ="_blank">https://github.com/indigo-dc/DEEPaaS</a>
**Version:** 0.5.2
