### 2019-06-21 #1517 @trallard
Hey @dangunter, @krother, @rougier thanks a lot for offering to review this submission.

Each of you have been assigned a checklist for your review, please go over the items and check them if you consider them to be satisfied.
Please use this issue for ongoing discussions. If you need to start a  longer discussion or request please make an issue on the submission repo and reference it here.

If you encounter any issues or have questions, please feel free to ping me here.

Thanks again for your help 🙏🏼
