### 2019-07-16 #1517 @rougier
@trallard Can we have a link to the paper at the top of the thread ? (I cannot find it)
