### 2019-07-17 #1517 @rougier
@alvarolopez I'm trying to test your software but I have problems because of the gevent package that cannot compile v1.4.0 on my machine (OSX). Last version seems to compile ok but not the 1.4.0. Do you know through which package this dependency comes from ?
