### 2019-07-17 #1517 @alvarolopez
@rougier this is a dependency on DEEPaaS itself, but we are depending on version 1.2.0 (actually I now realise that this can be bumped to 1.3.0).

What error are you obtaining?
