+++ requirements.txt
@@ -9,7 +9,7 @@ oslo.config>=2.3.0 # Apache-2.0
 
 flask-restplus>=0.12.1
 Flask>=1.0.2
-Werkzeug>=0.14.1
+Werkzeug>=0.15.0
 stevedore>=1.20.0 # Apache-2.0
 
 gevent==1.2.1
