### 2019-07-18 #1517 @rougier
Lots of warnings and errors which make it difficult to track the actual error. Among them:

```
  src/gevent/libev/gevent.corecext.c:30182:21: error: no member named 'exc_type' in 'struct _ts'
      *type = tstate->exc_type;
              ~~~~~~  ^
  src/gevent/libev/gevent.corecext.c:30183:22: error: no member named 'exc_value' in 'struct _ts'; did you mean 'curexc_value'?
      *value = tstate->exc_value;
                       ^~~~~~~~~
```
