### 2019-07-23 88c4cf0
add get_train/test_args docstrings
