+++ requirements.txt
@@ -12,4 +12,4 @@ Flask>=1.0.2
 Werkzeug>=0.15.0
 stevedore>=1.20.0 # Apache-2.0
 
-gevent==1.2.1
+gevent==1.4.0
