### 2019-07-23 3de0f4d
Prepare 0.5.0 release
