+++ releasenotes/notes/allow-passing-args-dbadafef832265e9.yaml
@@ -1,0 +1,7 @@
+---
+features:
+  - Allow to specify which arguments the underlying model accepts for the testing
+    calls. In this release users can implement a "get_test_args" function in the
+    module entry point, allowing to expose those arguments through the API.
+fixes:
+  - https://jira.deep-hybrid-datacloud.eu/browse/DPD-489
