### 2019-07-23 #1517 @alvarolopez
@rouger I have fixed this issue, related with an old pinned dependency version and Python3. Version 0.5.0 should work.
