### 2019-07-24 #1517 @rougier
I confirm it fixed my problem.
