### 2019-07-24 #1517 @rougier
@alvarolopez I think you're missing some DOI in the references of your paper. Also, I did not find any way to test the package. Would you that be possible to have a step by step description of a minimal test to run?
