### 2019-07-24 033d60a
update docs and docstrs
