+++ releasenotes/notes/allow-passing-args-dbadafef832265e9.yaml
@@ -5,3 +5,7 @@ features:
     module entry point, allowing to expose those arguments through the API.
 fixes:
   - https://jira.deep-hybrid-datacloud.eu/browse/DPD-489
+issues:
+  - The 0.5.0 version stops passing down the raw data when making a prediction,
+	and passes down the file objects. This breaks backwards compatibility with
+	the 0.4.0 version.
