### 2019-08-23 09f9c80
Fix test args parser
This is a patch to fig a bug in test parser that keep piling up the test
args of the previous modules.

For example if the module loading order is image-classification and then
satellites, the image classification predict method module would show
the test args of image-classification but the satellites predict method
would show the test args of both satellites and image-classification.
