### 2019-09-04 #1517 @alvarolopez
Hi @rougier, catching up after summer break.

I will correct the DOIs as soon as possible.

Regarding the instructions to run the API is to use the "make develop" command, documented [here](https://deepaas.readthedocs.io/en/stable/quickstart.html) I reckon that docs must be updated properly, with more detailed information. Currently there is more information on how to run it in the project-wide documentation [here](https://docs.deep-hybrid-datacloud.eu/en/latest/user/overview/api.html).

If you want to test a real model that is integrated with the API, you can have a loot at [here](https://docs.deep-hybrid-datacloud.eu/en/latest/user/howto/try-service-locally.html)
