### 2019-09-05 dc9533d
Fix documentation build error
It seems that the release notes caused some error when building the
documentation (due to tabs mixed with spaces). This change fixes it.
