### 2019-09-05 b1a81de
Added the option of choosing between image or json
