+++ deepaas/api/v1.py
@@ -14,8 +14,6 @@
 # License for the specific language governing permissions and limitations
 # under the License.
 
-import copy
-
 import flask
 import flask_restplus
 from flask_restplus import fields
@@ -55,6 +53,18 @@ data_parser.add_argument('url',
                          dest='urls',
                          required=False,
                          action="append")
+
+data_parser.add_argument('output',
+                         help="Output format (image or json)",
+                         type=str,
+                         default='json',
+                         choices=['json','image'],
+                         dest='output',
+                         required=False,
+                         action="append")
+
+
+
 
 model_links = api.model('Location', {
     "rel": fields.String(required=True),
@@ -165,12 +175,10 @@ for model_name, model_obj in model.MODELS.items():
             return m
 
     # Fill the test parser with the supported arguments. Different models may
-    # have different arguments. We get here a copy of the original parser,
-    # since otherwise if we have several models the arguments will pile up.
-    test_parser = copy.deepcopy(data_parser)
+    # have different arguments.
     test_args = model_obj.get_test_args()
     for k, v in test_args.items():
-        test_parser.add_argument(k, **v)
+        data_parser.add_argument(k, **v)
 
     @api.marshal_with(response, envelope='resource')
     @api.route('/%s/predict' % model_name)
@@ -178,12 +186,12 @@ for model_name, model_obj in model.MODELS.items():
         model_name = model_name
         model_obj = model_obj
 
-        @api.expect(test_parser)
+        @api.expect(data_parser)
         def post(self):
             """Make a prediction given the input data."""
 
-            args = test_parser.parse_args()
-
+            args = data_parser.parse_args()
+            print("args ---> ", args)
             if (not any([args["urls"], args["files"]]) or
                     all([args["urls"], args["files"]])):
                 raise exceptions.BadRequest("You must provide either 'url' or "
@@ -193,8 +201,8 @@ for model_name, model_obj in model.MODELS.items():
                 # FIXME(aloga): only handling one file, see comment on top of
                 # file and [1] for more details
                 # [1] https://github.com/noirbizarre/flask-restplus/issues/491
-                args["files"] = [args["files"]]
-
+                # data = [f.read() for f in args["files"]]
+                # data = [args["files"].read()]
                 ret = self.model_obj.predict_data(args)
             elif args["urls"]:
                 ret = self.model_obj.predict_url(args)
