### 2019-09-05 6778afb
Output can be image or json for inference
