### 2019-09-05 acd7faf
Output can be image or json for inference
