### 2019-09-09 bb90bb2
Choose between image or json as output
