### 2019-09-09 1c81970
Type of output passed as a header
