### 2019-09-09 b6883fb
Removed append from the output header request
