### 2019-09-16 88b7979
Correct bibliography with correct entries and add missing DOI items
