+++ paper/paper.bib
@@ -12,7 +12,6 @@
     year="2019",
     month="Jan",
     day="19",
-    abstract="The combined impact of new computing resources and techniques with an increasing avalanche of large datasets, is transforming many research areas and may lead to technological breakthroughs that can be used by billions of people. In the recent years, Machine Learning and especially its subfield Deep Learning have seen impressive advances. Techniques developed within these two fields are now able to analyze and learn from huge amounts of real world examples in a disparate formats. While the number of Machine Learning algorithms is extensive and growing, their implementations through frameworks and libraries is also extensive and growing too. The software development in this field is fast paced with a large number of open-source software coming from the academy, industry, start-ups or wider open-source communities. This survey presents a recent time-slide comprehensive overview with comparisons as well as trends in development and usage of cutting-edge Artificial Intelligence software. It also provides an overview of massive parallelism support that is capable of scaling computation effectively and efficiently in the era of Big Data.",
     issn="1573-7462",
     doi="10.1007/s10462-018-09679-z",
     url="https://doi.org/10.1007/s10462-018-09679-z"
@@ -27,15 +26,19 @@
  publisher = {University of California, Irvine},
 }
 
-@article{foster2005service,
-  title={Service-oriented science},
-  author={Foster, Ian},
-  journal={Science},
-  volume={308},
-  number={5723},
-  pages={814--817},
-  year={2005},
-  publisher={American Association for the Advancement of Science}
+@article {Foster814,
+  author = {Foster, Ian},
+  title = {Service-Oriented Science},
+  volume = {308},
+  number = {5723},
+  pages = {814--817},
+  year = {2005},
+  doi = {10.1126/science.1110411},
+  publisher = {American Association for the Advancement of Science},
+  issn = {0036-8075},
+  URL = {https://science.sciencemag.org/content/308/5723/814},
+  eprint = {https://science.sciencemag.org/content/308/5723/814.full.pdf},
+  journal = {Science}
 }
 
 @online{openapi,
@@ -52,13 +55,16 @@
 }
 
 @article{zorrilla2013service,
-  title={A service oriented architecture to provide data mining services for non-expert data miners},
-  author={Zorrilla, Marta and Garc{\'\i}a-Saiz, Diego},
-  journal={Decision Support Systems},
-  volume={55},
-  number={1},
-  pages={399--411},
-  year={2013},
-  publisher={Elsevier}
+  title = "A service oriented architecture to provide data mining services for non-expert data miners",
+  journal = "Decision Support Systems",
+  volume = "55",
+  number = "1",
+  pages = "399 - 411",
+  year = "2013",
+  issn = "0167-9236",
+  doi = "https://doi.org/10.1016/j.dss.2012.05.045",
+  url = "http://www.sciencedirect.com/science/article/pii/S016792361200156X",
+  author = "Marta Zorrilla and Diego García-Saiz",
+  keywords = "Analytics service, BI-as-a-Service, Knowledge discovery database, Data mining, Service-oriented architecture, Web Services",
 }
 
