### 2019-09-16 #1517 @alvarolopez
@trallard I've fixed the missing DOI in several publications, under the `paper` branch.
