### 2019-09-16 #1517 @rougier
Catching up. Would it be possible to have a (even minimal) set of unittests to check everything has been installed properly and is running as expected ?
