### 2019-09-17 #1517 @alvarolopez
@rougier unit tests do exist (https://github.com/indigo-dc/DEEPaaS/tree/master/deepaas/tests) with a coverage of about 80%. A full fledged testing is being done using Jenkins: https://jenkins.indigo-datacloud.eu/job/Pipeline-as-code/job/DEEPaaS/job/master/
