### 2019-09-17 0932b87
Install netifaces from APT repo
netifaces (dependency netifaces->oslo.log->oslo.utils) fails to build on
the Dockerfile, due to a dependency error (GCC is not installed in the
original image).
