### 2019-09-17 #1517 @rougier
@trallard @alvarolopez With that last point (unit tests) and the DOI correction I think the paper is ready to be published on my side.  Any news from @dangunter and @krother?
