### 2019-09-17 d49df98
Adjust tox.ini settings and environments
Remove Banding from being a pep8 test and add a check to ensure that we
can render the README.md for PyPI uploading. Also update tox.ini with
more modern configuration
