+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 2.1
-envlist = py{36,27},pep8,pip-missing-reqs
+envlist = py{36,27},pep8,pip-missing-reqs,bandit,pypi
 skipsdist = True
 
 [testenv]
@@ -66,23 +66,14 @@ commands =
 envdir = {toxworkdir}/shared
 commands =
   flake8
-  # Run security linter
-  # B110: except: pass
-  # B410: importing etree
-  bandit -r deepaas -x tests -s B110,B410
 
 [testenv:bandit]
 envdir = {toxworkdir}/shared
 commands = bandit -r deepaas -x tests -s B110,B410
 
-[flake8]
-# H803 skipped on purpose per list discussion.
-# E123, E125 skipped as they are invalid PEP-8.
-
-show-source = True
-ignore = E123,E125,H803,H405
-builtins = _
-exclude=.venv,.git,.tox,dist,doc,*lib/python*,*egg,build
+[testenv:pypi]
+envdir = {toxworkdir}/shared
+commands = python3 setup.py check -r -s -m
 
 [testenv:cobertura]
 commands = py.test --cov=deepaas --cov-report=xml --cov-report=term-missing deepaas/tests
@@ -111,3 +102,12 @@ commands =
 # of the requirements.txt files
 deps = pip_check_reqs
 commands=pip-missing-reqs -d --ignore-file=deepaas/tests/* deepaas
+
+[flake8]
+# H803 skipped on purpose per list discussion.
+# E123, E125 skipped as they are invalid PEP-8.
+
+show-source = True
+ignore = E123,E125,H803,H405
+builtins = _
+exclude=.venv,.git,.tox,dist,doc,*lib/python*,*egg,build
