### 2019-09-17 #1517 @alvarolopez
@rougier I finally managed to get the correct DOIs being rendered, you can check the proofs here: https://github.com/openjournals/joss-papers/blob/joss.01517/joss.01517/10.21105.joss.01517.pdf
