### 2019-09-23 #1517 @alvarolopez
@krother, @rougier many thanks.

@trallard , what are the next steps to be done?
