### 2019-09-24 #1517 @trallard
It seems like @dangunter has gone MIA but since we have the reviews from @krother  and @rougier if they are happy for me to proceed / recommend acceptance I am happy to move this forward

---

In the meantime @alvarolopez  can you perform the following actions:

1. Create a new release if needed and update us on the final version of the software
2. Check again the paper for typos, missing DOIs and the such
3. Create an archive  (on Zenodo, figshare, or other) and post the archive DOI on this issue.

Once this is completed we can move over 😄
