### 2019-09-25 #1517 @alvarolopez
@trallard ok, what I would do is:

1. Merge the branch `paper` into the `master` branch so that the paper is built correctly.
2. Prepare a new release
3. Upload the new version into https://digital.csic.es/handle/10261/173108, that is our institutional repository. I need to check if a new version can be added there or if I need a new entry.

I will keep you posted on this.
