### 2019-09-30 2731762
Fix bibiography entry in JOSS paper
