+++ paper/paper.bib
@@ -42,7 +42,7 @@
 }
 
 @online{openapi,
-  author = {OpenAPI Initiative},
+  author = {{OpenAPI Initiative}},
   title = {OpenAPI Specification},
   year = 2018,
   url = {https://github.com/OAI/OpenAPI-Specification}
