### 2019-10-02 #1517 @alvarolopez
@t

> @trallard ok, what I would do is:
>
>     1. Merge the branch `paper` into the `master` branch so that the paper is built correctly.
>
>     2. Prepare a new release
>
>     3. Upload the new version into https://digital.csic.es/handle/10261/173108, that is our institutional repository. I need to check if a new version can be added there or if I need a new entry.
>
>
> I will keep you posted on this.

@trallard I think we are ready to go. The paper in the master branch is updated, and the Zenodo record is here: https://doi.org/10.5281/zenodo.1544377

Anything else from our side?
