### 2019-10-11 c614d93
Refactor API namespaces loading, using Flask blueprints
In order to ease the management of several API applications and
namespaces, we start using Flask blueprints, and then we simply register
them at the Flask application directly.
