### 2019-10-11 2ea0f07
Add a version endpoint at the root URL
Since we are going to start managing different API versions, lets add an
API endpoint at the "/" url, with links to the different API versions
that are being defined.

Sem-Ver: api-break
