+++ deepaas/tests/test_v1_api.py
@@ -54,52 +54,52 @@ class TestApiV1(base.TestCase):
         self.assertIn(response.status_code, [200, 201])
 
     def test_not_found(self):
-        ret = self.app.get("/models/%s" % uuid.uuid4().hex)
+        ret = self.app.get("/v1/models/%s" % uuid.uuid4().hex)
         self.assertEqual(404, ret.status_code)
 
-        ret = self.app.put("/models/%s" % uuid.uuid4().hex)
+        ret = self.app.put("/v1/models/%s" % uuid.uuid4().hex)
         self.assertEqual(404, ret.status_code)
 
-        ret = self.app.post("/models/%s" % uuid.uuid4().hex)
+        ret = self.app.post("/v1/models/%s" % uuid.uuid4().hex)
         self.assertEqual(404, ret.status_code)
 
-        ret = self.app.delete("/models/%s" % uuid.uuid4().hex)
+        ret = self.app.delete("/v1/models/%s" % uuid.uuid4().hex)
         self.assertEqual(404, ret.status_code)
 
     def test_model_not_found(self):
-        ret = self.app.put("/models/%s/train" % uuid.uuid4().hex)
+        ret = self.app.put("/v1/models/%s/train" % uuid.uuid4().hex)
         self.assertEqual(404, ret.status_code)
 
-        ret = self.app.post("/models/%s/predict" % uuid.uuid4().hex)
+        ret = self.app.post("/v1/models/%s/predict" % uuid.uuid4().hex)
         self.assertEqual(404, ret.status_code)
 
-        ret = self.app.get("/models/%s" % uuid.uuid4().hex)
+        ret = self.app.get("/v1/models/%s" % uuid.uuid4().hex)
         self.assertEqual(404, ret.status_code)
 
     def test_train(self):
-        ret = self.app.put("/models/deepaas-test/train")
+        ret = self.app.put("/v1/models/deepaas-test/train")
         self.assertEqual(501, ret.status_code)
 
     def test_predict_not_data(self):
-        ret = self.app.post("/models/deepaas-test/predict")
+        ret = self.app.post("/v1/models/deepaas-test/predict")
         self.assertEqual(400, ret.status_code)
 
     def test_predict_data_not_implemented(self):
         f = six.BytesIO(b"foo")
         ret = self.app.post(
-            "/models/deepaas-test/predict",
+            "/v1/models/deepaas-test/predict",
             data={"data": (f, "foo.txt")})
         self.assertEqual(501, ret.status_code)
 
     def test_predict_urls_not_implemented(self):
         ret = self.app.post(
-            "/models/deepaas-test/predict",
+            "/v1/models/deepaas-test/predict",
             data={"url": "http://example.org/"})
         self.assertEqual(501, ret.status_code)
 
     def test_predict_various_urls_not_implemented(self):
         ret = self.app.post(
-            "/models/deepaas-test/predict",
+            "/v1/models/deepaas-test/predict",
             data={"url": ["http://example.org/", "http://example.com"]})
         self.assertEqual(501, ret.status_code)
 
@@ -111,7 +111,7 @@ class TestApiV1(base.TestCase):
         with mock.patch.object(deepaas.model, "MODELS", {"fake": m}):
             m.predict_data.return_value = {}
             ret = self.app.post(
-                "/models/fake/predict",
+                "/v1/models/fake/predict",
                 data={"data": (f, "foo.txt")})
             m.predict_data.assert_called_with([content])
             self.assertEqual(200, ret.status_code)
@@ -124,7 +124,7 @@ class TestApiV1(base.TestCase):
         with mock.patch.object(deepaas.model, "MODELS", {"fake": m}):
             m.predict_url.return_value = {}
             ret = self.app.post(
-                "/models/fake/predict",
[…]
