### 2019-10-11 2854fe6
Add V2 skeleton, following V1 API but on separate modules
Sem-Ver: api-break
