### 2019-10-11 a4fa5bd
Add V2 skeleton, following V1 API but on separate modules
Sem-Ver: api-break
