### 2019-10-15 f310279
Use release_string as version, as it includes VCS tags
