### 2019-10-15 ac44a0e
Disable loading of V1 version and add CLI option
Sem-Ver: api-break
Sem-Ver: deprecation
