+++ deepaas/api/v2/predict.py
@@ -24,7 +24,7 @@ import werkzeug.exceptions as exceptions
 from deepaas import model
 
 # Get the models (this is a singleton, so it is safe to call it multiple times
-model.register_models()
+model.register_v2_models()
 
 ns = flask_restplus.Namespace(
     'models',
@@ -93,7 +93,7 @@ response = ns.model('ModelResponse', {
 # Therefore, in the next lines we iterate over the loaded models and create
 # the different resources for each model. This way we can also load the
 # expected parameters if needed (as in the training method).
-for model_name, model_obj in model.MODELS.items():
+for model_name, model_obj in model.V2_MODELS.items():
     # Fill the test parser with the supported arguments. Different models may
     # have different arguments. We get here a copy of the original parser,
     # since otherwise if we have several models the arguments will pile up.
