+++ deepaas/model/v1.py
@@ -1,0 +1,282 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018 Spanish National Research Council (CSIC)
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+import abc
+
+from oslo_log import log
+import six
+import werkzeug.exceptions as exceptions
+
+from deepaas.model import loading
+
+LOG = log.getLogger(__name__)
+
+# Model registry
+MODELS = {}
+MODELS_LOADED = False
+
+
+def register_models():
+    global MODELS
+    global MODELS_LOADED
+
+    if MODELS_LOADED:
+        return
+
+    try:
+        for name, model in loading.get_available_models("v1").items():
+            MODELS[name] = ModelWrapper(name, model)
+    except Exception as e:
+        LOG.warning("Error loading models: %s", e)
+
+    if not MODELS:
+        LOG.info("No models found, loading test model")
+        MODELS["deepaas-test"] = ModelWrapper("deepaas-test", TestModel())
+    MODELS_LOADED = True
+
+
+@six.add_metaclass(abc.ABCMeta)
+class BaseModel(object):
+    """Base class for all models to be used with DEEPaaS.
+
+    Note that it is not needed for DEEPaaS to inherit from this abstract base
+    class in order to expose the model functionality, but the entrypoint that
+    is configured should expose the same API.
+    """
+
+    @abc.abstractmethod
+    def predict_file(self, path):
+        """Perform a prediction from a file in the local filesystem.
+
+        This method will perform a prediction based on a file stored in the
+        local filesystem.
+
+        :param str path: path to the file to be analized
+        """
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def predict_data(self, args):
+        """Perform a prediction from the data passed in the arguments.
+
+        :param dict args: This is a Python dict containing all the necessary
+        information to make a prediction. The keys of this dict are:
+
+            * 'files': This is a Werkzeug `FileStorage` object containing the
+            data to predict. You can access the bytes of the data with
+            args['files'].read() or the file extension with
+            args['files'].content_type
+            * 'urls': This should be empty for this method
+            * other keys needed to make the prediction as defined by the
+            function `get_test_args`
+
+        :return: The response can be a str, a dict or a file (using for example
+        `flask.send_file`)
+        """
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def predict_url(self, args):
+        """Perform a prediction from a remote URL.
+
+        :param dict args: This is a Python dict containing all the necessary
+        information to make a prediction. The keys of this dict are:
+
+            * 'files': This should be empty for this method
+            * 'urls': This is a list of strs of the URLs to use for prediction
+            * other keys needed to make the prediction as defined by the
+            function `get_test_args`
+
+        :return: The response can be a str, a dict or a file (using for example
+        `flask.send_file`)
+        """
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def get_metadata(self):
+        """Return metadata from the exposed model.
+
+        :return: dictionary containing the model's metadata.
+        """
+        raise NotImplementedError()
+
+    @abc.abstractmethod
+    def train(self, *args):
+        """TBD."""
[…]
