### 2019-10-15 #1517 @trallard
@alvarolopez thanks for your patience.
I have made 2 PRs to add some changes prior to publication, one is to fix the broken reference as per above and the other are minimal suggestions (can or not approve this one).

Please ping me when the DOI fix is merged to proceed.
