### 2019-10-16 b2a458a
Update paper.md
Added some changes for paper acceptance
