### 2019-10-16 #1517 @alvarolopez
@trallard ping :-)

Changes were merged, thank you for your support!
