### 2019-10-18 37412ee
tox: Include testing for Py3.7
