+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 2.1
-envlist = py{36,27},pep8,pip-missing-reqs,bandit,pypi
+envlist = py{36,37,27},pep8,pip-missing-reqs,bandit,pypi
 skipsdist = True
 
 [testenv]
@@ -49,9 +49,9 @@ commands =
   {[testenv]commands}
   stestr run {posargs} 
 
-[testenv:py35]
+[testenv:py37]
 # TODO(efried): Remove this once https://github.com/tox-dev/tox/issues/425 is fixed.
-basepython = python3.5
+basepython = python3.7
 commands =
   {[testenv]commands}
   stestr run {posargs}
@@ -60,7 +60,7 @@ commands =
 # TODO(efried): Remove this once https://github.com/tox-dev/tox/issues/425 is fixed.
 basepython = python3.6
 commands =
-  {[testenv:py35]commands}
+  {[testenv:py37]commands}
 
 [testenv:pep8]
 envdir = {toxworkdir}/shared
