### 2019-10-21 e2d6470
Include upgrade notes in documentation
