+++ doc/source/upgrade.rst
@@ -1,0 +1,9 @@
+.. _upgrade-notes:
+
+Upgrade notes
+=============
+
+Upgrading to version 1.0.0
+--------------------------
+
+TBD
