+++ doc/source/api.rst
@@ -1,8 +1,30 @@
 .. _api:
 
-API
-===
+Internal API
+============
 
 .. currentmodule:: deepaas
 
-    .. automodule:: deepaas
+V2 Models
+---------
+
+Base Model
+##########
+
+.. automodule:: deepaas.model.v2.base
+   :members:
+   :inherited-members:
+
+Model wrapper
+#############
+
+.. automodule:: deepaas.model.v2
+   :members:
+   :inherited-members:
+
+V2 test model
+#############
+
+.. automodule:: deepaas.model.v2.test
+   :members:
+   :inherited-members:
