+++ doc/source/contributing/index.rst
@@ -1,0 +1,10 @@
+================================
+How to contribute to DEEPaaS API
+================================
+
+
+.. toctree::
+   :maxdepth: 1
+
+   contributing
+   code_of_conduct

[…]
