### 2019-10-21 5ce1110
V2: Initial version of the new API
This change implementes the following changes:

- New entrypoint namespace.
- New functions to add arguments to the endpoints
- New documentation

Sem-Ver: api-break

Reorder documentation
