+++ deepaas/model/v2/__init__.py
@@ -1,0 +1,239 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018 Spanish National Research Council (CSIC)
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+import jsonschema
+import jsonschema.exceptions
+from oslo_log import log
+import werkzeug.exceptions as exceptions
+
+from deepaas.model import loading
+from deepaas.model.v2 import test
+
+LOG = log.getLogger(__name__)
+
+# Model registry
+MODELS = {}
+MODELS_LOADED = False
+
+
+def register_models():
+    global MODELS
+    global MODELS_LOADED
+
+    if MODELS_LOADED:
+        return
+
+    try:
+        for name, model in loading.get_available_models("v2").items():
+            MODELS[name] = ModelWrapper(name, model)
+    except Exception as e:
+        LOG.warning("Error loading models: %s", e)
+
+    if MODELS:
+        MODELS_LOADED = True
+        return
+
+    LOG.warning("No models found using V2 namespace, trying with V1. This is "
+                "DEPRECATED and it is done only to try to preserve backards "
+                "compatibility, but may lead to unexpected behaviour. You "
+                "should move to the new namespace as soon as possible. Please "
+                "refer to the documentation to get more details.")
+
+    try:
+        for name, model in loading.get_available_models("v1").items():
+            MODELS[name] = ModelWrapper(name, model)
+    except Exception as e:
+        LOG.warning("Error loading models: %s", e)
+
+    if not MODELS:
+        LOG.info("No models found with V2 or V1 namespace, loading test model")
+        MODELS["deepaas-test"] = ModelWrapper("deepaas-test", test.TestModel())
+    MODELS_LOADED = True
+
+
+class ModelWrapper(object):
+    """Class that will wrap the loaded models before exposing them.
+
+    Whenever a model is loaded it will be wrapped with this class to create a
+    wrapper object that will handle the calls to the model's methods so as to
+    handle non-existent method exceptions.
+
+    :param name: Model name
+    :param model: Model object
+    :raises exceptions.InternalServerError: in case that a model has defined
+        a reponse schema that is nod JSON schema valid (DRAFT 4)
+    """
+    def __init__(self, name, model):
+        self.name = name
+        self.model = model
+
+        schema = getattr(self.model, "response", None)
+
+        if schema is not None:
+            try:
+                jsonschema.Draft4Validator.check_schema(schema)
+            except jsonschema.exceptions.SchemaError as e:
+                LOG.exception(e)
+                raise exceptions.InternalServerError(
+                    description="Model defined schema is invalid, "
+                                "check server logs."
+                )
+            self.response_validator = jsonschema.Draft4Validator(schema)
+            self.has_schema = True
+        else:
+            self.response_validator = None
+            self.has_schema = False
+
+        self.response_schema = schema
+
+    def validate_response(self, response):
+        """Validate a response against the model's response schema, if set.
+
+        If the wrapped model has defined a ``response`` attribute we will
+        validate the response that
+
+        :param response: The reponse that will be validated.
+        :raises exceptions.InternalServerError: in case the reponse cannot be
+            validated.
+        """
+        if self.has_schema is not True:
+            raise exceptions.InternalServerError(
[…]
