+++ requirements.txt
@@ -13,3 +13,5 @@ Werkzeug>=0.15.0
 stevedore>=1.20.0 # Apache-2.0
 
 gevent==1.4.0
+
+jsonschema>=2.6.0
