+++ doc/source/user/index.rst
@@ -1,0 +1,10 @@
+User and developer documentation
+================================
+
+.. toctree::
+   :maxdepth: 1
+
+   quickstart
+   v1-api
+   v2-api
+   openwhisk
