### 2019-10-22 48c88e5
Add debug endpoint
