+++ releasenotes/notes/add-debug-endpoint-7b2f3ede199d74cc.yaml
@@ -1,0 +1,6 @@
+---
+features:
+  - New debug option, can be enabled with the ``--debug-endpoint``. This will
+    activate the ``/debug`` endpoint where all the output generated by the
+    model (i.e. using ``print``) or logged in the ``deepaas.api`` logger will
+    be shown.
