### 2019-10-22 4500ad8
Add debug endpoint
