### 2019-10-23 a983624
doc: update CLI man pages
