+++ doc/source/cli/index.rst
@@ -1,0 +1,8 @@
+Available commands
+==================
+
+
+.. toctree::
+   :maxdepth: 1
+
+   *
