### 2019-10-23 #1517 @trallard
Ok I think this review is ready to be sent over to @openjournals/joss-eics  for their editorial duties.
I am recommending acceptance now that the reviews and my editorial checks have been completed.
