### 2019-10-23 960284a
Disable reloading when running as debug
By default, Flask and werkzeug reload the application if they see that
there are changes in the app files. We disable this, even in debug, as
reloading a whole model can be troublesome in some situations.
