### 2019-10-24 #1517 @alvarolopez
@kthyng done, check https://doi.org/10.5281/zenodo.1544377
