### 2019-10-24 #1517 @kthyng
@alvarolopez Actually can you check again? The title in Zenodo should be "DEEPaaS API: a REST API for Machine Learning and Deep Learning models" right?
