### 2019-10-25 #1517 @kthyng
@alvarolopez Ok great, though the version is now inconsistent between this and Zenodo. Here, it is listed as 0.5.2, which is the latest release on github. Is that correct? If so, please update it on Zenodo to match.
