### 2019-10-25 #1517 @alvarolopez
@kthyng :man_facepalming: yes indeed. I forgot to create the release for that version.
