### 2019-10-25 #1517 @kthyng
Congrats @alvarolopez on your new paper!!

Many thanks to @krother and @rougier for your time and expertise reviewing, and to @trallard for editing!
