### 2019-10-25 #1517 @kthyng
@arfon I just realized that the final paper still lists as a reviewer someone who ended up not participated but is still listed. Can we remove them?
