### 2019-10-26 #1517 @arfon
> @arfon I just realized that the final paper still lists as a reviewer someone who ended up not participated but is still listed. Can we remove them?

Sure thing, I can fix that.
