### 2019-10-28 #1517 @alvarolopez
@trallard @kthyng @krother @rougier thanks for taking care of the whole process.

@arfon: I just realized that my surname and name are not correctly identified. My name is "Álvaro" and surname is "López García" (i.e. I have two surnames). Can this be fixed?

Many thanks in advance.
