### 2019-10-28 #1517 @arfon
> @arfon: I just realized that my surname and name are not correctly identified. My name is "Álvaro" and surname is "López García" (i.e. I have two surnames). Can this be fixed?

Sorry about that @alvarolopez. That should be fixed now. If the PDF on the JOSS site is still looking out of date it might take a while to refresh as there's caching in place.
